backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq
sentence-transformers
torch
faiss-cpu
requests
python-dotenv
